Used a software from AMD: http://developer.amd.com/tools/shader/Pages/default.aspx#download This software allow to use shaders and simulate a compilation for AMD/ATI hardware. With it I was able to fix the shaders. The water and terrain shaders should run now on ATI hardware. (Rumors says that ATI is stricter on the syntax)
git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@248 528fbfa9-60f6-4f4a-a82b-d133dc65ca33

+++ media/shaders/ocean.frag
@@ -10,7 +10,7 @@ uniform int terrainScale;
 
 void main()
 {
-	int scale = terrainScale * 0.145;
+	float scale = terrainScale * 0.145f;
 	vec2 texCoord = vec2(gl_TexCoord[0]);
 	
 	vec4 tex0    = texture2D( oceanNormalMap, vec2(texCoord.x*5 + waterTime,texCoord.y*3));
