+++ media/shaders/ocean.vert
@@ -3,7 +3,7 @@ uniform int terrainScale;
 
 void main(void)
 {
-	int scale = (terrainScale * 0.05);
+	float scale = (terrainScale * 0.05f);
 			
 	posY = gl_Vertex.y;
